+++ lib/src/main/java/com/metaplex/lib/programs/token_metadata/gpa_builders/TokenGpaBuilder.kt
@@ -15,7 +15,7 @@ class TokenGpaBuilder(
 
     fun whereMint(mint: PublicKey): TokenGpaBuilder {
         val mutableGpaBulder = this
-        return mutableGpaBulder.where(0.toUInt(), mint)
+        return mutableGpaBulder.where(0, mint)
     }
 
     fun selectOwner(): TokenGpaBuilder {
@@ -25,7 +25,7 @@ class TokenGpaBuilder(
 
     fun whereOwner(owner: PublicKey): TokenGpaBuilder {
         val mutableGpaBulder = this
-        return mutableGpaBulder.where(32.toUInt(), owner)
+        return mutableGpaBulder.where(32, owner)
     }
 
     fun selectAmount(): TokenGpaBuilder {
@@ -35,22 +35,22 @@ class TokenGpaBuilder(
 
     fun whereAmount(amount: Int): TokenGpaBuilder {
         val mutableGpaBulder = this
-        return mutableGpaBulder.where(64.toUInt(), amount)
+        return mutableGpaBulder.where(64, amount)
     }
 
     fun whereDoesntHaveDelegate(): TokenGpaBuilder {
         val mutableGpaBulder = this
-        return mutableGpaBulder.where(72.toUInt(), 0.toByte())
+        return mutableGpaBulder.where(72, 0.toByte())
     }
 
     fun whereHasDelegate(): TokenGpaBuilder {
         var mutableGpaBulder = this
-        return mutableGpaBulder.where(72.toUInt(), 1.toByte())
+        return mutableGpaBulder.where(72, 1.toByte())
     }
 
     fun whereDelegate(delegate: PublicKey): TokenGpaBuilder {
         var mutableGpaBulder = this
         mutableGpaBulder = mutableGpaBulder.whereHasDelegate()
-        return mutableGpaBulder.where(76.toUInt(), delegate)
+        return mutableGpaBulder.where(76, delegate)
     }
 }
